merge Post

+++ src/main/java/com/sparta/springnewsfeed/service/PostService.java
@@ -1,9 +1,13 @@
 package com.sparta.springnewsfeed.service;
 
+import com.sparta.springnewsfeed.exception.InvalidCredentialsException;
 import com.sparta.springnewsfeed.dto.PostRequestDto;
 import com.sparta.springnewsfeed.dto.PostResponseDto;
 import com.sparta.springnewsfeed.entity.Post;
+import com.sparta.springnewsfeed.entity.User;
 import com.sparta.springnewsfeed.repository.PostRepository;
+import com.sparta.springnewsfeed.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -12,9 +16,11 @@ import java.util.List;
 @Service
 public class PostService {
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    public PostService(PostRepository postRepository) {
+    public PostService(PostRepository postRepository, UserRepository userRepository) {
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     public PostResponseDto findById(Long postId) {
@@ -24,33 +30,46 @@ public class PostService {
         return new PostResponseDto(post);
     }
 
-    public PostResponseDto createPost(PostRequestDto requestDto) {
+    public PostResponseDto createPost(PostRequestDto requestDto, Long userId) {
         Post post = new Post(requestDto);
+        User user = userRepository.findById(userId).orElseThrow(
+                () -> new NullPointerException("삭제된 유저 입니다.")
+        );
+        post.setUser(user);
         Post savePost = postRepository.save(post);
-
+        user.getPosts().add(savePost);
+        userRepository.save(user);
         return new PostResponseDto(savePost);
     }
 
-    public PostResponseDto updateById(Long postId, PostRequestDto requestDto) {
+    public PostResponseDto updateById(Long postId, PostRequestDto requestDto, Long userId) {
         Post post = postRepository.findById(postId).orElseThrow(
                 () -> new NullPointerException("해당 게시물이 존재하지 않습니다.")
         );
+        if (!post.getUser().getId().equals(userId)) {
+            throw new InvalidCredentialsException("권한이 없습니다.");
+        }
         post.setTitle(requestDto.getTitle());
         post.setContent(requestDto.getContent());
         Post savePost = postRepository.save(post);
         return new PostResponseDto(savePost);
     }
 
-    public void deleteById(Long postId) {
+    public void deleteById(Long postId, Long userId) {
         Post post = postRepository.findById(postId).orElseThrow(
-                () -> new NullPointerException("해당 게시물이 존재하지 않습니다.")
+                () -> new EntityNotFoundException("해당 게시물이 존재하지 않습니다.")
         );
-        postRepository.delete(post);
+        User user = post.getUser();
+        if (user.getId().equals(userId)) {
+            user.getPosts().remove(post);
+            postRepository.delete(post);
+        } else {
+            throw new InvalidCredentialsException("권한이 없습니다.");
+        }
     }
 
-    public List<PostResponseDto> findMyPost() {
-        Long userid = 123L;
-        List<Post> myPosts = postRepository.findByUserId(userid).orElseThrow(
+    public List<PostResponseDto> findMyPost(Long userId) {
+        List<Post> myPosts = postRepository.findByUserId(userId).orElseThrow(
                 () -> new NullPointerException("작성한 게시물이 없습니다.")
         );
         List<PostResponseDto> postResponseDtos = new ArrayList<>();
@@ -60,4 +79,5 @@ public class PostService {
         }
         return postResponseDtos;
     }
+
 }
